Update to include ConfigMap deletion when not needed and update tests.
