+++ Dockerfile
@@ -1,5 +1,5 @@
 # Build the manager binary
-FROM golang:1.22 AS builder
+FROM golang:1.22.0 AS builder
 ARG TARGETOS
 ARG TARGETARCH
