+++ internal/modelcontroller/files.go
@@ -1,11 +1,84 @@
 package modelcontroller
 
 import (
+	"context"
+	"fmt"
+	"reflect"
 	"strings"
 
 	kubeaiv1 "github.com/substratusai/kubeai/api/v1"
 	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/log"
 )
+
+// getModelFilesConfigMapName returns the name of the ConfigMap for a model
+func getModelFilesConfigMapName(model *kubeaiv1.Model) string {
+	return fmt.Sprintf("model-%s-files", model.Name)
+}
+
+// ensureModelFilesConfigMap ensures that the ConfigMap for model files exists and is up to date
+func (r *ModelReconciler) ensureModelFilesConfigMap(ctx context.Context, model *kubeaiv1.Model) error {
+	log := log.FromContext(ctx)
+	configMapName := getModelFilesConfigMapName(model)
+
+	// Build the expected data map with sanitized keys
+	expectedData := make(map[string]string)
+	for _, file := range model.Spec.Files {
+		sanitizedKey := fileConfigMapKey(file.Path)
+		expectedData[sanitizedKey] = file.Content
+	}
+
+	// Check if ConfigMap exists
+	existingConfigMap := &corev1.ConfigMap{}
+	err := r.Get(ctx, client.ObjectKey{Namespace: model.Namespace, Name: configMapName}, existingConfigMap)
+
+	if err != nil {
+		if !apierrors.IsNotFound(err) {
+			return fmt.Errorf("getting model files configmap: %w", err)
+		}
+
+		// ConfigMap doesn't exist, create it
+		configMap := &corev1.ConfigMap{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      configMapName,
+				Namespace: model.Namespace,
+			},
+			Data: expectedData,
+		}
+
+		// Set owner reference
+		if err := ctrl.SetControllerReference(model, configMap, r.Scheme); err != nil {
+			return fmt.Errorf("setting controller reference on model files configmap: %w", err)
+		}
+
+		if err := r.Create(ctx, configMap); err != nil {
+			return fmt.Errorf("creating model files configmap: %w", err)
+		}
+
+		log.Info("Created model files ConfigMap", "configMapName", configMapName)
+		return nil
+	} else if len(model.Spec.Files) == 0 {
+		if err := r.Delete(ctx, existingConfigMap); err != nil {
+			return fmt.Errorf("deleting empty model files configmap: %w", err)
+		}
+		return nil
+	}
+
+	// ConfigMap exists, check if update is needed
+	if !reflect.DeepEqual(existingConfigMap.Data, expectedData) {
+		existingConfigMap.Data = expectedData
+		if err := r.Update(ctx, existingConfigMap); err != nil {
+			return fmt.Errorf("updating model files configmap: %w", err)
+		}
+		log.Info("Updated model files ConfigMap", "configMapName", configMapName)
+	}
+
+	return nil
+}
 
 // patchFileVolumes adds volume and volume mount for the model files if specified.
 func patchFileVolumes(spec *corev1.PodSpec, model *kubeaiv1.Model) {
