+++ internal/modelcontroller/model_controller.go
@@ -29,7 +29,6 @@ import (
 	"k8s.io/client-go/rest"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -77,70 +76,6 @@ type ModelReconciler struct {
 //+kubebuilder:rbac:groups="",resources=pods/status,verbs=get;update;patch
 //+kubebuilder:rbac:groups="",resources=pods/finalizers,verbs=update
 //+kubebuilder:rbac:groups="",resources=configmaps,verbs=get;list;watch;create;update;patch;delete
-
-// getModelFilesConfigMapName returns the name of the ConfigMap for a model
-func getModelFilesConfigMapName(model *kubeaiv1.Model) string {
-	return fmt.Sprintf("model-%s-files", model.Name)
-}
-
-// ensureModelFilesConfigMap ensures that the ConfigMap for model files exists and is up to date
-func (r *ModelReconciler) ensureModelFilesConfigMap(ctx context.Context, model *kubeaiv1.Model) error {
-	if len(model.Spec.Files) == 0 {
-		return nil
-	}
-
-	log := log.FromContext(ctx)
-	configMapName := getModelFilesConfigMapName(model)
-
-	// Build the expected data map with sanitized keys
-	expectedData := make(map[string]string)
-	for _, file := range model.Spec.Files {
-		sanitizedKey := fileConfigMapKey(file.Path)
-		expectedData[sanitizedKey] = file.Content
-	}
-
-	// Check if ConfigMap exists
-	existingConfigMap := &corev1.ConfigMap{}
-	err := r.Get(ctx, client.ObjectKey{Namespace: model.Namespace, Name: configMapName}, existingConfigMap)
-
-	if err != nil {
-		if !apierrors.IsNotFound(err) {
-			return fmt.Errorf("getting model files configmap: %w", err)
-		}
-
-		// ConfigMap doesn't exist, create it
-		configMap := &corev1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      configMapName,
-				Namespace: model.Namespace,
-			},
-			Data: expectedData,
-		}
-
-		// Set owner reference
-		if err := ctrl.SetControllerReference(model, configMap, r.Scheme); err != nil {
-			return fmt.Errorf("setting controller reference on model files configmap: %w", err)
-		}
-
-		if err := r.Create(ctx, configMap); err != nil {
-			return fmt.Errorf("creating model files configmap: %w", err)
-		}
-
-		log.Info("Created model files ConfigMap", "configMapName", configMapName)
-		return nil
-	}
-
-	// ConfigMap exists, check if update is needed
-	if !reflect.DeepEqual(existingConfigMap.Data, expectedData) {
-		existingConfigMap.Data = expectedData
-		if err := r.Update(ctx, existingConfigMap); err != nil {
-			return fmt.Errorf("updating model files configmap: %w", err)
-		}
-		log.Info("Updated model files ConfigMap", "configMapName", configMapName)
-	}
-
-	return nil
-}
 
 func (r *ModelReconciler) Reconcile(ctx context.Context, req ctrl.Request) (res ctrl.Result, resErr error) {
 	log := log.FromContext(ctx)
